Exclude Gifs
It seems like the exceptGif is not working. When we use this package gifs still show up although they are greyed out. When they are imported our app crashes. 

`exceptGif` is deprecated in the library. PR is switching to the replacement `exceptMimeType(listOf(MimeType.GIF))` 

This could be expanded in the future to give users the option to specify the formats they'd like to exclude.